scripts/pdf_to_context/ocr_service/requirements.txt
# DeepSeek-OCR Microservice - зависимости

# vLLM для inference
vllm>=0.3.0  # vLLM engine

# FastAPI для HTTP API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Обработка изображений
Pillow>=10.0.0

# DeepSeek-OCR модель (загружается через huggingface)
# Требует trust_remote_code=True в vLLM

# Дополнительно для продакшена
# gunicorn>=21.2.0  # Production ASGI server
# prometheus-client>=0.19.0  # Метрики

